@remotion/studio: Reorganize render modal
