+++ packages/studio/src/components/RenderModal/ServerRenderModal.tsx
@@ -45,6 +45,7 @@ import {FileIcon} from '../../icons/file';
 import {PicIcon} from '../../icons/frame';
 import {GearIcon} from '../../icons/gear';
 import {GifIcon} from '../../icons/gif';
+import {FilmIcon} from '../../icons/video';
 import {ModalsContext} from '../../state/modals';
 import {SidebarContext} from '../../state/sidebar';
 import {Button} from '../Button';
@@ -86,8 +87,9 @@ import type {RenderType} from './RenderModalAdvanced';
 import {RenderModalAdvanced} from './RenderModalAdvanced';
 import {RenderModalAudio} from './RenderModalAudio';
 import {RenderModalBasic} from './RenderModalBasic';
+import {RenderModalEncoding, type QualityControl} from './RenderModalEncoding';
+import {RenderModalEnvironmentVariables} from './RenderModalEnvironmentVariables';
 import {RenderModalGif} from './RenderModalGif';
-import type {QualityControl} from './RenderModalPicture';
 import {RenderModalPicture} from './RenderModalPicture';
 import {
 	ResolveCompositionBeforeModal,
@@ -1484,6 +1486,30 @@ const RenderModal: React.FC<
 							GIF
 						</VerticalTab>
 					) : null}
+					{shownTabs.includes('encoding') ? (
+						<VerticalTab
+							style={horizontalTab}
+							selected={tab === 'encoding'}
+							onClick={() => setTab('encoding')}
+						>
+							<div style={iconContainer}>
+								<FilmIcon style={icon} color="currentcolor" />
+							</div>
+							Encoding
+						</VerticalTab>
+					) : null}
+					{shownTabs.includes('environment') ? (
+						<VerticalTab
+							style={horizontalTab}
+							selected={tab === 'environment'}
+							onClick={() => setTab('environment')}
+						>
+							<div style={iconContainer}>
+								<DataIcon style={icon} />
+							</div>
+							Environment
+						</VerticalTab>
+					) : null}
 					{shownTabs.includes('advanced') ? (
 						<VerticalTab
 							style={horizontalTab}
@@ -1527,30 +1553,11 @@ const RenderModal: React.FC<
 							renderMode={renderMode}
 							scale={scale}
 							setScale={setScale}
-							pixelFormat={pixelFormat}
-							pixelFormatOptions={pixelFormatOptions}
 							imageFormatOptions={imageFormatOptions}
-							crf={crf}
-							setCrf={setCrf}
-							customTargetVideoBitrate={customTargetVideoBitrate}
-							maxCrf={maxCrf}
-							minCrf={minCrf}
 							jpegQuality={jpegQuality}
-							qualityControlType={qualityControlType}
 							setJpegQuality={setJpegQuality}
-							setColorSpace={setColorSpace}
-							colorSpace={colorSpace}
-							setCustomTargetVideoBitrateValue={
-								setCustomTargetVideoBitrateValue
-							}
-							setQualityControl={setQualityControl}
 							videoImageFormat={videoImageFormat}
 							stillImageFormat={stillImageFormat}
-							shouldDisplayQualityControlPicker={supportsBothQualityControls}
-							encodingBufferSize={encodingBufferSize}
-							setEncodingBufferSize={setEncodingBufferSize}
-							encodingMaxRate={encodingMaxRate}
-							setEncodingMaxRate={setEncodingMaxRate}
 							compositionWidth={resolvedComposition.width}
 							compositionHeight={resolvedComposition.height}
 						/>
@@ -1591,6 +1598,43 @@ const RenderModal: React.FC<
 							setLimitNumberOfGifLoops={setLimitNumberOfGifLoops}
 							setNumberOfGifLoopsSetting={setNumberOfGifLoopsSetting}
 						/>
+					) : tab === 'encoding' ? (
+						<RenderModalEncoding
+							renderMode={renderMode}
+							codec={codec}
+							qualityControlType={qualityControlType}
+							setQualityControl={setQualityControl}
+							shouldDisplayQualityControlPicker={supportsBothQualityControls}
+							crf={crf}
+							setCrf={setCrf}
+							maxCrf={maxCrf}
+							minCrf={minCrf}
+							customTargetVideoBitrate={customTargetVideoBitrate}
+							setCustomTargetVideoBitrateValue={
+								setCustomTargetVideoBitrateValue
+							}
+							encodingBufferSize={encodingBufferSize}
+							setEncodingBufferSize={setEncodingBufferSize}
+							encodingMaxRate={encodingMaxRate}
+							setEncodingMaxRate={setEncodingMaxRate}
+							pixelFormat={pixelFormat}
+							pixelFormatOptions={pixelFormatOptions}
+							colorSpace={colorSpace}
+							setColorSpace={setColorSpace}
+							x264Preset={x264Preset}
+							setx264Preset={setx264Preset}
+							gopSize={gopSize}
+							setGopSize={setGopSize}
+							hardwareAcceleration={hardwareAcceleration}
+							setHardwareAcceleration={setHardwareAcceleration}
+							disallowParallelEncoding={disallowParallelEncoding}
+							setDisallowParallelEncoding={setDisallowParallelEncoding}
+						/>
+					) : tab === 'environment' ? (
+						<RenderModalEnvironmentVariables
+							envVariables={envVariables}
+							setEnvVariables={setEnvVariables}
+						/>
 					) : tab === 'data' ? (
 						<DataEditor
 							defaultProps={inputProps}
@@ -1605,10 +1649,6 @@ const RenderModal: React.FC<
 						/>
 					) : (
 						<RenderModalAdvanced
-							x264Preset={x264Preset}
-							setx264Preset={setx264Preset}
-							gopSize={gopSize}
-							setGopSize={setGopSize}
 							concurrency={concurrency}
 							maxConcurrency={maxConcurrency}
 							minConcurrency={minConcurrency}
@@ -1616,8 +1656,6 @@ const RenderModal: React.FC<
 							setConcurrency={setConcurrency}
 							delayRenderTimeout={delayRenderTimeout}
 							setDelayRenderTimeout={setDelayRenderTimeout}
-							disallowParallelEncoding={disallowParallelEncoding}
-							setDisallowParallelEncoding={setDisallowParallelEncoding}
 							setDisableWebSecurity={setDisableWebSecurity}
 							setIgnoreCertificateErrors={setIgnoreCertificateErrors}
 							setHeadless={setHeadless}
@@ -1626,8 +1664,6 @@ const RenderModal: React.FC<
 							disableWebSecurity={disableWebSecurity}
 							openGlOption={openGlOption}
 							setOpenGlOption={setOpenGlOption}
-							setEnvVariables={setEnvVariables}
-							envVariables={envVariables}
 							offthreadVideoCacheSizeInBytes={offthreadVideoCacheSizeInBytes}
 							setMediaCacheSizeInBytes={setMediaCacheSizeInBytes}
 							mediaCacheSizeInBytes={mediaCacheSizeInBytes}
@@ -1638,15 +1674,12 @@ const RenderModal: React.FC<
 							setOffthreadVideoThreads={setOffthreadVideoThreads}
 							enableMultiProcessOnLinux={multiProcessOnLinux}
 							setChromiumMultiProcessOnLinux={setChromiumMultiProcessOnLinux}
-							codec={codec}
 							userAgent={userAgent}
 							setUserAgent={setUserAgent}
 							setBeep={setBeepOnFinish}
 							beep={beepOnFinish}
 							repro={repro}
 							setRepro={setRepro}
-							hardwareAcceleration={hardwareAcceleration}
-							setHardwareAcceleration={setHardwareAcceleration}
 							chromeModeOption={chromeMode}
 							setChromeModeOption={setChromeMode}
 							darkMode={darkMode}
